adds roles admin | user

+++ src/modules/catalog/resolvers/equipmentType.resolver.ts
@@ -5,7 +5,13 @@ import { EquipmentTypeService } from '../service/equipmentType.service';
 import { formatZodErrors } from '../../../utils/errors';
 
 export const Query = {
-  equipmentTypes: async (_: unknown, __: unknown, { db }: Context) => {
+  equipmentTypes: async (
+    _: unknown,
+    __: unknown,
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new EquipmentTypeService(db).getEquipmentTypes();
   },
   equipmentType: async (
@@ -19,8 +25,13 @@ export const Mutation = {
   createEquipmentType: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateEquipmentTypeInputSchema.parse(input);
 
@@ -42,8 +53,13 @@ export const Mutation = {
   deleteEquipmentType: async (
     _: unknown,
     { id }: { id: string },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new EquipmentTypeService(db).deleteEquipmentType(id);
   },
 };
