+++ src/modules/catalog/resolvers/metrologyControlType.resolver.ts
@@ -5,7 +5,13 @@ import { MetrologyControlTypeService } from '../service/metrologyControlType.sev
 import { CreateMetrologyControlTypeInputSchema } from '../dto/CreateMetrologyControlTypeDto';
 
 export const Query = {
-  metrologyControlTypes: async (_: unknown, __: unknown, { db }: Context) => {
+  metrologyControlTypes: async (
+    _: unknown,
+    __: unknown,
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new MetrologyControlTypeService(db).getMetrologyControlTypes();
   },
   metrologyControlType: async (
@@ -19,8 +25,13 @@ export const Mutation = {
   createMetrologyControlType: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateMetrologyControlTypeInputSchema.parse(input);
 
@@ -42,8 +53,13 @@ export const Mutation = {
   deleteMetrologyControlType: async (
     _: unknown,
     { id }: { id: string },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new MetrologyControlTypeService(db).deleteMetrologyControlType(
       id
     );
