+++ src/modules/catalog/resolvers/scope.resolver.ts
@@ -5,7 +5,9 @@ import { ScopeService } from '../service/scope.service';
 import { CreateScopeInputSchema } from '../dto/CreateScopeDto';
 
 export const Query = {
-  scopes: async (_: unknown, __: unknown, { db }: Context) => {
+  scopes: async (_: unknown, __: unknown, { db, currentUser }: Context) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new ScopeService(db).getScopes();
   },
   scope: async (_: unknown, { id }: { id: string }, { db }: Context) => {},
@@ -15,8 +17,13 @@ export const Mutation = {
   createScope: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateScopeInputSchema.parse(input);
 
@@ -33,7 +40,16 @@ export const Mutation = {
     { id, input }: { id: string; input: unknown },
     { db }: Context
   ) => {},
-  deleteScope: async (_: unknown, { id }: { id: string }, { db }: Context) => {
+  deleteScope: async (
+    _: unknown,
+    { id }: { id: string },
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new ScopeService(db).deleteScope(id);
   },
 };
