+++ src/modules/catalog/resolvers/status.resolver.ts
@@ -5,7 +5,9 @@ import { StatusService } from '../service/status.service';
 import { CreateStatusInputSchema } from '../dto/CreateStatusDto';
 
 export const Query = {
-  statuses: async (_: unknown, __: unknown, { db }: Context) => {
+  statuses: async (_: unknown, __: unknown, { db, currentUser }: Context) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new StatusService(db).getAllStatuses();
   },
   status: async (_: unknown, { id }: { id: string }, { db }: Context) => {},
@@ -15,8 +17,13 @@ export const Mutation = {
   createStatus: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateStatusInputSchema.parse(input);
 
@@ -33,7 +40,16 @@ export const Mutation = {
     { id, input }: { id: string; input: unknown },
     { db }: Context
   ) => {},
-  deleteStatus: async (_: unknown, { id }: { id: string }, { db }: Context) => {
+  deleteStatus: async (
+    _: unknown,
+    { id }: { id: string },
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new StatusService(db).deleteStatus(id);
   },
 };
