+++ src/modules/device/resolvers/device.resolver.ts
@@ -3,20 +3,33 @@ import { ZodError } from 'zod';
 import { formatZodErrors } from '../../../utils/errors';
 
 import { Context } from '../../../context';
-import { DeviceEntity } from '../types/device.types';
 import { CreateDeviceInputSchema } from '../dto/CreateDeviceDto';
 import { DeviceService } from '../service/device.service';
 import { UpdateDeviceInputSchema } from '../dto/UpdateDeviceDto';
 
 export const Query = {
-  devices: async (_: unknown, __: unknown, { db }: Context) => {
+  devices: async (_: unknown, __: unknown, { db, currentUser }: Context) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new DeviceService(db).getDevices();
   },
 
-  devicesWithRelations: async (_: unknown, __: unknown, { db }: Context) => {
+  devicesWithRelations: async (
+    _: unknown,
+    __: unknown,
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new DeviceService(db).getDevicesWithRelations();
   },
-  device: async (_: unknown, { id }: { id: string }, { db }: Context) => {
+  device: async (
+    _: unknown,
+    { id }: { id: string },
+    { db, currentUser }: Context
+  ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
     return await new DeviceService(db).getDevice(id);
   },
 };
@@ -25,8 +38,13 @@ export const Mutation = {
   createDevice: async (
     _: unknown,
     { input }: { input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateDeviceInputSchema.parse(input);
       return await new DeviceService(db).createDevice(validatedInput);
@@ -55,8 +73,13 @@ export const Mutation = {
   updateDevice: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ) => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = UpdateDeviceInputSchema.parse(input);
       return await new DeviceService(db).updateDevice(id, validatedInput);
@@ -70,8 +93,13 @@ export const Mutation = {
   deleteDevice: async (
     _: unknown,
     { id }: { id: string },
-    { db }: Context
+    { db, currentUser }: Context
   ): Promise<boolean> => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new DeviceService(db).deleteDevice(id);
   },
 };
