+++ src/modules/user/resolvers/user.resolver.ts
@@ -7,14 +7,24 @@ import { Context } from '../../../context';
 import { CreateUserInputSchema } from '../dto/CreateUserDto';
 
 export const Query = {
-  users: async (_: unknown, __: unknown, { db }: Context): Promise<User[]> => {
+  users: async (
+    _: unknown,
+    __: unknown,
+    { db, currentUser }: Context
+  ): Promise<User[]> => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new UserService(db).getUsers();
   },
   user: async (
     _: unknown,
     { id }: { id: string },
-    { db }: Context
+    { db, currentUser }: Context
   ): Promise<User> => {
+    if (!currentUser) throw new Error('Не авторизован');
     return await new UserService(db).getUser(id);
   },
 };
@@ -23,8 +33,13 @@ export const Mutation = {
   updateUser: async (
     _: unknown,
     { id, input }: { id: string; input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ): Promise<User> => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = UpdateUserInputSchema.parse(input);
       return await new UserService(db).updateUser(id, validatedInput);
@@ -38,8 +53,13 @@ export const Mutation = {
   createUser: async (
     _: unknown,
     { input }: { input: unknown },
-    { db }: Context
+    { db, currentUser }: Context
   ): Promise<User> => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     try {
       const validatedInput = CreateUserInputSchema.parse(input);
       return await new UserService(db).createUser(validatedInput);
@@ -53,8 +73,13 @@ export const Mutation = {
   deleteUser: async (
     _: unknown,
     { id }: { id: string },
-    { db }: Context
+    { db, currentUser }: Context
   ): Promise<boolean> => {
+    if (!currentUser) throw new Error('Не авторизован');
+
+    if (currentUser.role !== 'admin') {
+      throw new Error('Доступ запрещен: нужны права администратора');
+    }
     return await new UserService(db).deleteUser(id);
   },
 };
